requirements.txt
Chameleon==2.11
Mako==0.8.1
MarkupSafe==0.18
PasteDeploy==1.5.0
SQLAlchemy==0.8.2
WebOb==1.2.3
py==1.4.15
pyramid==1.4.3
pytest==2.3.5
repoze.filesafe==2.0b1
repoze.lru==0.6
six==1.3.0
testfixtures==3.0.1
translationstring==1.1
venusian==1.0a8
zope.deprecation==4.0.2
zope.interface==4.0.5